Refactor positioning logic to custom hook
